requirements.txt
srsly==2.4.1
matplotlib==3.2.1
numpy==1.21.3
pandas==1.3.2
pandas_datareader==0.10.0
plotly==5.1.0
scipy==1.7.1
seaborn==0.11.1
session_info==1.0.0
yfinance==0.1.64
IPython==7.28.0
jupyter_client==7.0.6
jupyter_core==4.8.1
jupyterlab==3.1.18
notebook==6.4.4
streamlit==1.1.0
yfinance==0.1.64
sklearn==0.0
python-dateutil==2.8.2
tiingo==0.14.0